backend/requirements.txt
FastAPI
gunicorn
line-bot-sdk
langchain
langchain-openai
uvicorn
pyowm
langchain-google-genai
pillow
langchain_google_vertexai
langchain-google-memorystore-redis
langchain-google-firestore
firebase_admin
pysen[lint]
google-cloud-bigquery
